Jorge

+++ src/main/java/com/example/proyectotalentotech/services/UsuarioService.java
@@ -32,7 +32,12 @@ public class UsuarioService {
         return usuarioRepository.save(usuario);
     }
 
+    @Transactional
     public void eliminar(Integer id) {
-        usuarioRepository.deleteById(id);
+        if (usuarioRepository.existsById(id)) {
+            usuarioRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("El usuario con ID " + id + " no existe.");
+        }
     }
 }
